split map file into 7 files
