+++ source/views/map-carls/urls.js
@@ -1,0 +1,7 @@
+// @flow
+
+export const MAP_DATA_URL = 'https://carls-app.github.io/map-data/map.json'
+export const CARLETON_TILE_TEMPLATE =
+    'https://www.carleton.edu/global_stock/images/campus_map/tiles/base/{z}_{x}_{y}.png'
+export const GITHUB_TILE_TEMPLATE =
+    'https://carls-app.github.io/map-data/tiles/{z}/{x}_{y}.png'
